Committing progress
Added curlscripts to the project on curl-scripts/auth/all of the files
as well as curl-scripts/recipes/all of the files. In api.js i made
changed to better route the the requests to the api because they
were not set up correctly. in events.js I edited the submit handler
callbacks better where I had mistakes. in ui.js i created better
handlers for the responses. Modified index.html because my forms were
not reading the information properly.. Site still doesnt function
the way I want it to.. It recieved all the respones and handles the
events well for the auth sections. but my recipes section is still not
working properly.