Comitting additions
Added information to my readme.md such as the user stories and
also added my wireFrame. To create and update I added the functionality
to add and edit the preperation time for the recipe. In index.html I
added the form and input to put in the prepTime initially and to edit.