Committing a working application
Fixed all errors && achieved a working application. I had a missing "p"
in the word exports in my UI.js. after fixing that error, I ran into
other errors which I fixed. now, I can add a recipe, edit recipes, view
1 or many recipes. My next steps will be to display this information to
the user in order for the user to be able to communicate with the app
and edit themselves after log-in.