Committing styling additions
To app.js I added functionality to hide and show certain fields
which I added to index.html. in /auth/ui.js I added the functionality to
show and hide certain things when not needed, also added a template to
help the end-user create a recipe with ease. in recipe/ui.js I finilized
and made the functionality work that allows users to see the recipe they
created as well as other recipes that are currently on the system by
other users. in index.scss I added stying to the page using flexbox and
in index.html I organized things a bit better to make it easier to
control the flow of the page.