Final Commit Functioning app
This is the final commit for a functioning application. In ui.js I've
eddited the files to ommit any unneeded console.logs and made sure the
logic works properly. in api.js the same. in events.js I added a
if statement to the see one recipe to make sure they cannot send an
invalid ID. in index.html I changed the "token" words to the word "ID"